+++ pystore/config.py
@@ -4,7 +4,7 @@
 # PyStore: Flat-file datastore for timeseries data
 # https://github.com/ranaroussi/pystore
 #
-# Copyright 2018-2019 Ran Aroussi
+# Copyright 2018-2020 Ran Aroussi
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,4 +19,11 @@
 # limitations under the License.
 
 from os.path import expanduser as _expanduser
+
 DEFAULT_PATH = _expanduser("~/pystore")
+DEFAULT_PARTITION_SIZE = 99e+6  # ~99MB
+PARTITION_SIZE = 99e+6  # ~99MB
+
+# dask distributed
+_SCHEDULER = None
+_CLIENT = None