+++ pystore/utils.py
@@ -4,7 +4,7 @@
 # PyStore: Flat-file datastore for timeseries data
 # https://github.com/ranaroussi/pystore
 #
-# Copyright 2018-2019 Ran Aroussi
+# Copyright 2018-2020 Ran Aroussi
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -25,6 +25,8 @@ import shutil
 import pandas as pd
 import numpy as np
 from dask import dataframe as dd
+from dask.distributed import Client
+
 
 try:
     from pathlib import Path
@@ -151,3 +153,32 @@ def delete_store(store):
 def delete_stores():
     shutil.rmtree(get_path())
     return True
+
+
+def set_client(scheduler=None):
+    if scheduler != config._SCHEDULER and config._CLIENT is not None:
+        try:
+            config._CLIENT.shutdown()
+        except Exception:
+            pass
+
+    if scheduler is not None:
+        config._SCHEDULER = scheduler
+        config._CLIENT = Client(scheduler)
+
+    return config._CLIENT
+
+
+def get_client():
+    return config._CLIENT
+
+
+def set_partition_size(size=None):
+    if size is None:
+        size = config.DEFAULT_PARTITION_SIZE * 1
+    config.PARTITION_SIZE = size
+    return config.PARTITION_SIZE
+
+
+def get_partition_size():
+    return config.PARTITION_SIZE